Add admin test routes/views
