+++ app/controllers/admin/tests_controller.rb
@@ -1,0 +1,53 @@
+class Admin::TestsController < ApplicationController
+  before_action :find_test, only: %i[show edit update destroy start]
+
+  def index
+    @tests = Test.all
+  end
+
+  def show; end
+
+  def new
+    @test = Test.new
+  end
+
+  def create
+    @test = Test.new(test_params)
+
+    if @test.save
+      redirect_to @test
+    else
+      render :new
+    end
+  end
+
+  def update
+    @test = Test.find(params[:id])
+
+    if @test.save
+      redirect_to @test
+    else
+      render :edit
+    end
+  end
+
+  def destroy
+    @test.destroy
+    redirect_to admin_tests_path
+  end
+
+  def start
+    current_user.tests.push(@test)
+    redirect_to current_user.test_passage(@test)
+  end
+
+  private
+
+  def find_test
+    @test = Test.find(params[:id])
+  end
+
+  def test_params
+    params.require(:test).permit(:title, :level, :category_id)
+  end
+end
