+++ app/controllers/application_controller.rb
@@ -4,4 +4,9 @@ class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
 
   before_action :authenticate_user!
+
+  def after_sign_in_path_for(resource)
+    pp resource
+    resource.admin? ? admin_tests_path : super
+  end
 end
