+++ app/controllers/tests_controller.rb
@@ -1,48 +1,18 @@
 class TestsController < ApplicationController
-  before_action :find_test, only: %i[show edit update destroy start]
+  before_action :find_test, only: %i[start]
 
   def index
     @tests = Test.all
   end
-
-  def new
-    @test = Test.new
-  end
-
-  def create
-    @test = Test.new(test_params)
-
-    if @test.save
-      redirect_to @test
-    else
-      render :new
-    end
-  end
-
-  def show; end
 
   def start
     current_user.tests.push(@test)
     redirect_to current_user.test_passage(@test)
   end
 
-  def update
-    @test = Test.find(params[:id])
-
-    if @test.save
-      redirect_to @test
-    else
-      render :edit
-    end
-  end
-
   private
 
   def find_test
     @test = Test.find(params[:id])
-  end
-
-  def test_params
-    params.require(:test).permit(:title, :level, :category_id)
   end
 end
