+++ app/helpers/admin/tests_helper.rb
@@ -1,0 +1,2 @@
+module Admin::TestsHelper
+end
