+++ app/models/user.rb
@@ -26,4 +26,8 @@ class User < ApplicationRecord
   def test_passage(test)
     test_passages.order(created_at: :desc).find_by(test_id: test.id)
   end
+
+  def admin?
+    is_a?(Admin)
+  end
 end
