+++ app/views/admin/questions/_question.html.erb
@@ -1,0 +1,6 @@
+<tr>
+  <td><%= question.body %></td>
+  <td><%= link_to 'Show', admin_question_path(question) %></td>
+  <td><%= link_to 'Edit', edit_admin_question_path(question) %></td>
+  <td><%= link_to 'Delete', admin_question_path(question), method: :delete, data: { confirm: 'Are you sure?' } %></td>
+</tr>
