+++ app/views/admin/questions/index.html.erb
@@ -8,4 +8,4 @@
   <%= render @test.questions %>
 </table>
 
-<%= link_to 'Go Back', :root %> | <%= link_to 'Add new question', new_test_question_path %>
+<%= link_to 'Go Back', :root %> | <%= link_to 'Add new question', new_admin_test_question_path %>
