+++ app/views/admin/questions/show.html.erb
