+++ app/views/admin/tests/_test.html.erb
@@ -1,0 +1,12 @@
+<tr>
+  <td><%= test.title %></td>
+  <td><%= test.level %></td>
+  <td><%= test.questions.count %></td>
+  <td><%= link_to 'Show', admin_test_path(test) %></td>
+  <td><%= link_to 'Edit', edit_admin_test_path(test) %></td>
+  <td>
+    <%= link_to 'Delete', admin_test_path(test),
+                method: :delete,
+                data: { confirm: 'Are you sure?' } %>
+  </td>
+</tr>
