+++ app/views/admin/tests/index.html.erb
@@ -4,9 +4,10 @@
     <th>Title</th>
     <th>Level</th>
     <th>Questions</th>
-    <th>Action</th>
+    <th colspan="3">Action</th>
   </tr>
 
   <%= render @tests %>
 </table>
 <br>
+<%= link_to 'Add New Test', new_admin_test_path %>
