+++ app/views/admin/tests/show.html.erb
@@ -7,6 +7,6 @@
   </tr>
   <%= render @test.questions %>
 </table><br>
-<%= link_to 'Create new question', new_test_question_path(@test) %>
+<%= link_to 'Create new question', new_admin_test_question_path(@test) %>
 <br>
 <%= link_to 'Back', tests_path %>
