--- app/views/questions/_question.html.erb
@@ -1,6 +1,0 @@
-<tr>
-  <td><%= question.body %></td>
-  <td><%= link_to 'Show', question_path(question) %></td>
-  <td><%= link_to 'Edit', edit_question_path(question) %></td>
-  <td><%= link_to 'Delete', question_path(question), method: :delete, data: { confirm: 'Are you sure?' } %></td>
-</tr>
