+++ app/views/tests/_test.html.erb
@@ -3,11 +3,4 @@
   <td><%= test.level %></td>
   <td><%= test.questions.count %></td>
   <td><%= button_to 'Start', start_test_path(test) %></td>
-  <td><%= link_to 'Show', test_path(test) %></td>
-  <td><%= link_to 'Edit', edit_test_path(test) %></td>
-  <td>
-    <%= link_to 'Delete', test_path(test),
-                method: :delete,
-                data: { confirm: 'Are you sure?' } %>
-  </td>
 </tr>
