+++ test/controllers/admin/tests_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Admin::TestsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
